pyproject.toml
[tool.poetry]
name = "ira"
version = "3.0.0"
description = "Ira – Machinecraft AI Pantheon"
authors = ["Machinecraft"]
packages = [{ include = "ira", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0"
pydantic-settings = "^2.0"
httpx = ">=0.27"
tiktoken = ">=0.7"
qdrant-client = "^1.9"
neo4j = "^5.0"
sqlalchemy = {version = "^2.0", extras = ["asyncio"]}
aiosqlite = ">=0.20"
apscheduler = "^3.10"
flashrank = ">=0.2"
greenlet = "^3.3.2"
google-api-python-client = "^2.0"
google-auth-httplib2 = ">=0.2"
google-auth-oauthlib = "^1.0"
typer = ">=0.12"
rich = ">=13.0"
fastapi = ">=0.115"
uvicorn = {version = ">=0.30", extras = ["standard"]}
python-multipart = ">=0.0.9"
jinja2 = ">=3.1"
python-telegram-bot = ">=21.0"
pypdf = "^6.7.5"
openpyxl = "^3.1.5"
python-docx = "^1.2.0"

[tool.poetry.scripts]
ira = "ira.interfaces.cli:app"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = ">=0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
